+++ changelog.md
@@ -764,3 +764,4 @@ See https://github.com/Daveiano/weewx-wdc/compare/v3.3.0...580071ca175a03fe4924b
 - Updated windGust display in forecast table GH-240
 - Updated sorting of min/max graphs in the `temp_min_max_avg` graph GH-247
 - Added new property for charts: `legendPosition` - Change legend position GH-228
+- Added option to set default theme (light/dark/auto) GH-241